requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
black==22.3.0
cachetools==4.2.2
certifi==2021.5.30
click==8.1.3
colorama==0.4.4
ephem==4.0.0.2
Faker==8.14.1
flake8==4.0.1
greenlet==1.1.2
mccabe==0.6.1
more-itertools==8.9.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
PySocks==1.7.1
pytest==3.7.1
pytest-sugar==0.9.1
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.25
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2021.1
tzlocal==3.0
